Implement single line output.
There has been a bit of refactoring and moving functions around in new
files. This is not a bad thing.

As a side effect of this change, if the last character read from the
host is not \n, we output a \n. This was for single line output, to
make sure that each host resulted in at least oneoutput line, but is
also useful in normal mode.

Signed-off-by: Russell Steicke <russells@adelie.cx>